// lib/models/enums/common.ts
export enum ESolutionOrientation {
  Bottom = "Bottom",
  Left = "Left",
  Right = "Right",
}

export enum EPaymentPeriod {
  Monthly = "month",
  Yearly = "year",
  Lifetime = "lifetime",
  IndividualItems = "individualItems",
}

export enum ESubscriptionStatusType {
  Subscribed = "SUBSCRIBED",
  Default = "SUBSCRIBE",
  Downgrade = "DOWNGRADE",
}